@@ -84,7 +84,7 @@ public class SubjectParser implements Serializable {
                 // So we're getting the principals from the Subject here. Do note that we miss the
                 // potential extra deployment roles here which may be in the principals collection we get
                 // passed in.
-                Subject subject = (Subject) PolicyContext.getContext(JACC.SUBJECT_CONTAINER_KEY);
+                Subject subject = PolicyContext.getContext(JACC.SUBJECT_CONTAINER_KEY);
 
                 if (subject == null) {
                     return null;
@@ -322,8 +322,7 @@ public class SubjectParser implements Serializable {
 
             return true;
 
-        } catch (ClassNotFoundException | NoSuchMethodException | SecurityException | IllegalAccessException | IllegalArgumentException
-                | InvocationTargetException e) {
+        } catch (ClassNotFoundException | NoSuchMethodException | SecurityException | IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
             return false;
         }
     }
